+++ server/internal/handlers/replies.go
@@ -3,9 +3,11 @@ package handlers
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 	"net/http"
 	"time"
 
+	"github.com/golang-jwt/jwt/v5"
 	"github.com/google/uuid"
 )
 
@@ -41,7 +43,18 @@ func (h *APIHandler) CreateReply(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, "failed to decode reply", http.StatusBadRequest)
 		return
 	}
+	claims, ok := r.Context().Value("claims").(jwt.MapClaims)
+	if !ok {
+		http.Error(w, "no claims", http.StatusUnauthorized)
+		return
+	}
 
+	sub := claims["sub"].(string)
+	fmt.Println(sub)
+	if sub == "" {
+		http.Error(w, "no sub", http.StatusUnauthorized)
+		return
+	}
 	uid := r.PathValue("uid")
 	var err error
 	ans.QuestionUID, err = uuid.Parse(uid)
@@ -51,7 +64,7 @@ func (h *APIHandler) CreateReply(w http.ResponseWriter, r *http.Request) {
 	}
 	ctx, cancel := context.WithTimeout(r.Context(), 10*time.Second)
 	defer cancel()
-	_, err = h.DB.Exec(ctx, "insert into answers (content, question_uid) values ($1, $2)", ans.Content, ans.QuestionUID)
+	_, err = h.DB.Exec(ctx, "insert into answers (content, question_uid, author) values ($1, $2, $3)", ans.Content, ans.QuestionUID, sub)
 	if err != nil {
 		http.Error(w, "failed to save reply to db", http.StatusInternalServerError)
 		return
@@ -65,10 +78,20 @@ func (h *APIHandler) DeleteReply(w http.ResponseWriter, r *http.Request) {
 	defer r.Body.Close()
 	quid := r.PathValue("quid")
 	ruid := r.PathValue("ruid")
-
+	claims, ok := r.Context().Value("claims").(jwt.MapClaims)
+	if !ok {
+		http.Error(w, "no claims", http.StatusUnauthorized)
+		return
+	}
+	sub := claims["sub"].(string)
+	fmt.Println(sub)
+	if sub == "" {
+		http.Error(w, "no sub", http.StatusUnauthorized)
+		return
+	}
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
 	defer cancel()
-	_, err := h.DB.Exec(ctx, "delete from answers where uid = $1 and question_uid = $2", ruid, quid)
+	_, err := h.DB.Exec(ctx, "delete from answers where uid = $1 and question_uid = $2 and author = $3", ruid, quid, sub)
 	if err != nil {
 		http.Error(w, "failed to delete reply", http.StatusInternalServerError)
 		return
